MoreOOP Checkpoint 2 - Minimal requirements done
