+++ tasklogs/tasks.txt
@@ -1,5 +1,5 @@
 [T] | 0 | change amp
-[D] | 0 | remaster drums | tuesday 10am
 [E] | 1 | write guitar part | saturday 9pm
 [E] | 1 | live gig | saturday 9pm
 [T] | 0 | logos/sublunary
+[T] | 1 | baby blue
